Cmake branch updated to compile on my linux debian install

+++ controller/lib/src/linux/net_interface_imp.h
@@ -95,7 +95,7 @@ namespace avdecc_lib
         /**
          * Get network interface description by index.
          */
-        char * STDCALL get_dev_desc_by_index(uint32_t dev_index);
+        char * STDCALL get_dev_desc_by_index(size_t dev_index);
 
         /**
          * Select network interface by number.
